# apps/api-fastapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nha Tro Manager API",
    "description": "H\u1ec7 th\u1ed1ng RESTful API Qu\u1ea3n l\u00fd Nh\u00e0 tr\u1ecd theo ki\u1ebfn tr\u00fac T\u00e1ch bi\u1ec7t FastAPI & React SPA",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tenants": {
      "get": {
        "tags": [
          "Tenants"
        ],
        "summary": "Get Tenants",
        "operationId": "get_tenants_api_v1_tenants_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "T\u00ecm ki\u1ebfm theo t\u00ean, S\u0110T, s\u1ed1 CCCD",
              "title": "Search"
            },
            "description": "T\u00ecm ki\u1ebfm theo t\u00ean, S\u0110T, s\u1ed1 CCCD"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Pagesize"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_PaginatedData_TenantResponseSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tenants"
        ],
        "summary": "Create Tenant",
        "operationId": "create_tenant_api_v1_tenants_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantCreateSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_TenantResponseSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tenants/{tenant_id}/account": {
      "post": {
        "tags": [
          "Tenant Accounts"
        ],
        "summary": "Create Tenant Account",
        "operationId": "create_tenant_account_api_v1_tenants__tenant_id__account_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_TenantAccountCreateResponseSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_id}/account/reset-password": {
      "post": {
        "tags": [
          "Tenant Accounts"
        ],
        "summary": "Reset Tenant Password",
        "operationId": "reset_tenant_password_api_v1_tenants__tenant_id__account_reset_password_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_TempPasswordResponseSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_id}/account/suspend": {
      "post": {
        "tags": [
          "Tenant Accounts"
        ],
        "summary": "Suspend Tenant Account",
        "operationId": "suspend_tenant_account_api_v1_tenants__tenant_id__account_suspend_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_id}/account/reactivate": {
      "post": {
        "tags": [
          "Tenant Accounts"
        ],
        "summary": "Reactivate Tenant Account",
        "operationId": "reactivate_tenant_account_api_v1_tenants__tenant_id__account_reactivate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_PaginatedData_TenantResponseSchema__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginatedData_TenantResponseSchema_"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "requestId": {
            "type": "string",
            "title": "Requestid"
          }
        },
        "type": "object",
        "title": "APIResponse[PaginatedData[TenantResponseSchema]]"
      },
...
      "APIResponse_TempPasswordResponseSchema_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TempPasswordResponseSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "requestId": {
            "type": "string",
            "title": "Requestid"
          }
        },
        "type": "object",
        "title": "APIResponse[TempPasswordResponseSchema]"
      },
      "APIResponse_TenantAccountCreateResponseSchema_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TenantAccountCreateResponseSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "requestId": {
            "type": "string",
            "title": "Requestid"
          }
        },
        "type": "object",
        "title": "APIResponse[TenantAccountCreateResponseSchema]"
      },
      "APIResponse_TenantResponseSchema_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TenantResponseSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "requestId": {
            "type": "string",
            "title": "Requestid"
          }
        },
        "type": "object",
        "title": "APIResponse[TenantResponseSchema]"
      },
...
      "APIResponse_dict_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "requestId": {
            "type": "string",
            "title": "Requestid"
          }
        },
        "type": "object",
        "title": "APIResponse[dict]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedData_TenantResponseSchema_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TenantResponseSchema"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "pageSize": {
            "type": "integer",
            "title": "Pagesize"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "totalPages": {
            "type": "integer",
            "title": "Totalpages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "totalPages"
        ],
        "title": "PaginatedData[TenantResponseSchema]"
      },
...
      "TempPasswordResponseSchema": {
        "properties": {
          "tenantId": {
            "type": "string",
            "title": "Tenantid"
          },
          "userId": {
            "type": "string",
            "title": "Userid"
          },
          "tempPassword": {
            "type": "string",
            "title": "Temppassword"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "\u0110\u1eb7t l\u1ea1i m\u1eadt kh\u1ea9u th\u00e0nh c\u00f4ng. T\u1ea5t c\u1ea3 c\u00e1c phi\u00ean l\u00e0m vi\u1ec7c c\u0169 \u0111\u00e3 b\u1ecb h\u1ee7y."
          }
        },
        "type": "object",
        "required": [
          "tenantId",
          "userId",
          "tempPassword"
        ],
        "title": "TempPasswordResponseSchema"
      },
      "TenantAccountCreateResponseSchema": {
        "properties": {
          "tenantId": {
            "type": "string",
            "title": "Tenantid"
          },
          "userId": {
            "type": "string",
            "title": "Userid"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "tempPassword": {
            "type": "string",
            "title": "Temppassword"
          },
          "mustChangePassword": {
            "type": "boolean",
            "title": "Mustchangepassword",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "T\u1ea1o t\u00e0i kho\u1ea3n th\u00e0nh c\u00f4ng. M\u1eadt kh\u1ea9u t\u1ea1m th\u1eddi ch\u1ec9 hi\u1ec3n th\u1ecb m\u1ed9t l\u1ea7n duy nh\u1ea5t n\u00e0y."
          }
        },
        "type": "object",
        "required": [
          "tenantId",
          "userId",
          "phone",
          "tempPassword"
        ],
        "title": "TenantAccountCreateResponseSchema"
      },
      "TenantCreateSchema": {
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 1,
            "title": "Fullname",
            "description": "H\u1ecd v\u00e0 t\u00ean kh\u00e1ch thu\u00ea"
          },
          "phone": {
            "type": "string",
            "maxLength": 15,
            "minLength": 10,
            "title": "Phone",
            "description": "S\u1ed1 \u0111i\u1ec7n tho\u1ea1i"
          },
          "idCardNumber": {
            "type": "string",
            "maxLength": 20,
            "minLength": 9,
            "title": "Idcardnumber",
            "description": "S\u1ed1 CCCD/CMND"
          },
          "idCardIssuedDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idcardissueddate"
          },
          "idCardIssuedPlace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idcardissuedplace"
          },
          "dateOfBirth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dateofbirth"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "permanentAddress": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permanentaddress"
          },
          "vehicleNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehiclenumber"
          },
          "emergencyContactName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergencycontactname"
          },
          "emergencyContactPhone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergencycontactphone"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "fullName",
          "phone",
          "idCardNumber"
        ],
        "title": "TenantCreateSchema"
      },
      "TenantResponseSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "ownerId": {
            "type": "string",
            "title": "Ownerid"
          },
          "userId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Userid"
          },
          "fullName": {
            "type": "string",
            "title": "Fullname"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "idCardNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idcardnumber"
          },
          "idCardIssuedDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idcardissueddate"
          },
          "idCardIssuedPlace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idcardissuedplace"
          },
          "dateOfBirth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dateofbirth"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "permanentAddress": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permanentaddress"
          },
          "vehicleNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehiclenumber"
          },
          "emergencyContactName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergencycontactname"
          },
          "emergencyContactPhone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergencycontactphone"
          },
          "hasAccount": {
            "type": "boolean",
            "title": "Hasaccount",
            "default": false
          },
          "isAccountActive": {
            "type": "boolean",
            "title": "Isaccountactive",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ownerId",
          "fullName",
          "createdAt",
          "updatedAt"
        ],
        "title": "TenantResponseSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}